requirements.txt
beautifulsoup4==4.12.2
markdown2==2.4.10
python-dotenv==1.0.0
requests==2.31.0
atlassian-python-api==3.41.1
azure-search-documents>=11.4.0 